Add spinner component
