+++ src/components/Main.js
@@ -1,15 +1,18 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
+import RegisterContainer from '../container/RegisterContainer';
+import LoginContainer from '../container/LoginContainer';
 
 import WorkoutContainer from '../container/WorkoutContainer';
-import LoginOrSignup from './LoginOrSignup';
+import Spinner from './Spinner';
 
 const Main = () => {
   const token = useSelector((state) => state.token);
+  const loading = useSelector((state) => state.loading);
 
-  if (!token) {
-    return <LoginOrSignup token={token} />;
+  if (loading) {
+    return <Spinner />;
   }
   return (
     <Switch>
@@ -22,7 +25,18 @@ const Main = () => {
             : <Redirect to="/workouts" />
         )}
       />
-      <Route exact path="/workouts" component={WorkoutContainer} />
+      <Route
+        exact
+        path="/workouts"
+        render={() => (
+          !token
+            ? <Redirect to="/login" />
+            : <WorkoutContainer />
+        )}
+      />
+      <Route exact path="/login" component={LoginContainer} />
+      <Route exact path="/signup" component={RegisterContainer} />
+
     </Switch>
   );
 };
